requirements.txt
flask
gunicorn
sentence-transformers
faiss-cpu
numpy